# api/openapi.yaml
openapi: 3.0.3
info:
  title: Serenity ERP API
  description: |
    HIPAA-compliant API for Serenity Care Partners ERP system.
    
    ## Security
    - All endpoints require authentication via Bearer token
    - Role-based access control (RBAC) with attribute-based scoping (ABAC)
    - Row-level security enforced at database level
    - PHI endpoints require additional HIPAA compliance headers
    
    ## Rate Limiting
    - 1000 requests per minute per user
    - 10000 requests per minute per organization
    
    ## Data Classification
    Responses include `X-Data-Classification` header:
    - `public`: No sensitive information
    - `internal`: Internal business data
    - `confidential`: Sensitive business data
    - `phi`: Protected Health Information
    
  version: 1.0.0
  contact:
    name: Serenity Care Partners IT
    email: it@serenitycarepartners.com
  license:
    name: Proprietary
servers:
  - url: https://api.serenityerp.com/v1
    description: Production server
  - url: https://api-staging.serenityerp.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "123e4567-e89b-12d3-a456-426614174000"
    Timestamp:
      type: string
      format: date-time
      example: "2024-01-15T10:30:00Z"
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: object
      required: [error, message]
...
    User:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        organizationId:
          $ref: '#/components/schemas/UUID'
        email:
          type: string
          format: email
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        role:
          type: string
          enum: [founder, security_officer, compliance_officer, finance_director, billing_manager, rcm_analyst, scheduler, field_supervisor, hr_manager, credentialing_specialist, it_admin, support_agent, caregiver, client, family, payer_auditor, ai_service]
        phone:
          type: string
          maxLength: 20
        isActive:
          type: boolean
        lastLogin:
          $ref: '#/components/schemas/Timestamp'
        mfaEnabled:
          type: boolean
        hourlyRate:
          type: number
          format: decimal
          multipleOf: 0.01
        preferences:
          type: object
        createdAt:
          $ref: '#/components/schemas/Timestamp'
        updatedAt:
          $ref: '#/components/schemas/Timestamp'
      required: [id, organizationId, email, firstName, lastName, role, isActive, mfaEnabled, createdAt, updatedAt]
...
    EVVRecord:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        shiftId:
          $ref: '#/components/schemas/UUID'
        organizationId:
          $ref: '#/components/schemas/UUID'
        caregiverId:
          $ref: '#/components/schemas/UUID'
        clientId:
          $ref: '#/components/schemas/UUID'
        servicePerformed:
          type: string
          maxLength: 255
        clockInTime:
          $ref: '#/components/schemas/Timestamp'
        clockOutTime:
          $ref: '#/components/schemas/Timestamp'
        locationIn:
          type: object
          properties:
            latitude:
              type: number
              format: double
            longitude:
              type: number
              format: double
            accuracy:
              type: number
        locationOut:
          type: object
          properties:
            latitude:
              type: number
              format: double
            longitude:
              type: number
              format: double
            accuracy:
              type: number
        verificationMethod:
          type: string
          enum: [gps, telephony, fixed_device]
        submittedToSandata:
          type: boolean
        sandataTransactionId:
          type: string
          maxLength: 100
        isValid:
          type: boolean
        validationErrors:
          type: object
        createdAt:
          $ref: '#/components/schemas/Timestamp'
      required: [id, shiftId, caregiverId, clientId, servicePerformed, clockInTime, locationIn, verificationMethod, isValid, createdAt]
    ClockInRequest:
      type: object
      properties:
        shiftId:
          $ref: '#/components/schemas/UUID'
        servicePerformed:
          type: string
          maxLength: 255
        location:
          type: object
          properties:
            latitude:
              type: number
              format: double
            longitude:
              type: number
              format: double
            accuracy:
              type: number
          required: [latitude, longitude]
        verificationMethod:
          type: string
          enum: [gps, telephony, fixed_device]
        deviceInfo:
          type: object
      required: [shiftId, servicePerformed, location, verificationMethod]
      
    # Billing schemas
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                mfaCode:
                  type: string
                  maxLength: 6
              required: [email, password]
      responses:
        '200':
          description: Login successful
          headers:
            X-Data-Classification:
              schema:
                type: string
              example: internal
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
                  expiresAt:
                    $ref: '#/components/schemas/Timestamp'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '423':
          description: Account locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags: [Authentication]
      summary: User logout
      responses:
        '200':
          description: Logout successful
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /evv/clock-in:
    post:
      tags: [EVV]
      summary: Clock in for shift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClockInRequest'
      responses:
        '201':
          description: Clock in successful
          headers:
            X-Data-Classification:
              schema:
                type: string
              example: phi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EVVRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Already clocked in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /evv/clock-out:
    post:
      tags: [EVV]
      summary: Clock out from shift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                evvRecordId:
                  $ref: '#/components/schemas/UUID'
                location:
                  type: object
                  properties:
                    latitude:
                      type: number
                      format: double
                    longitude:
                      type: number
                      format: double
                    accuracy:
                      type: number
                  required: [latitude, longitude]
              required: [evvRecordId, location]
      responses:
        '200':
          description: Clock out successful
          headers:
            X-Data-Classification:
              schema:
                type: string
              example: phi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EVVRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
          
  # Claims endpoints
...